+++ src/main/java/com/vonage/client/messages/mms/MmsRequest.java
@@ -15,6 +15,7 @@
  */
 package com.vonage.client.messages.mms;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.vonage.client.messages.Channel;
 import com.vonage.client.messages.MessageRequest;
 import com.vonage.client.messages.MessageType;
@@ -22,22 +23,65 @@ import com.vonage.client.messages.internal.MessagePayload;
 
 public abstract class MmsRequest extends MessageRequest {
 	MessagePayload payload;
+	final Integer ttl;
 
 	protected MmsRequest(Builder<?, ?> builder, MessageType messageType) {
 		super(builder, Channel.MMS, messageType);
+		payload = new MessagePayload(builder.url, builder.caption);
+		payload.validateCaptionLength(2000);
+		int min = 300, max = 259200;
+		if ((this.ttl = builder.ttl) != null && (ttl < min || ttl > max)) {
+			throw new IllegalArgumentException("TTL must be between "+min+" and "+max+" seconds.");
+		}
+	}
+
+	@JsonProperty("ttl")
+	public Integer getTtl() {
+		return ttl;
 	}
 
 	@SuppressWarnings("unchecked")
 	protected abstract static class Builder<M extends MmsRequest, B extends Builder<? extends M, ? extends B>> extends MessageRequest.Builder<M, B> {
 		String url, caption;
+		Integer ttl;
 
-		protected B url(String url) {
+		/**
+		 * (REQUIRED)
+		 * Sets the media URL.
+		 *
+		 * @param url The URL as a string.
+		 * @return This builder.
+		 */
+		public B url(String url) {
 			this.url = url;
 			return (B) this;
 		}
 
+		/**
+		 * (OPTIONAL)
+		 * Additional text to accompany the media. Must be between 1 and 2000 characters.
+		 *
+		 * @param caption The caption string.
+		 * @return This builder.
+		 */
 		protected B caption(String caption) {
 			this.caption = caption;
+			return (B) this;
+		}
+
+		/**
+		 * (OPTIONAL)
+		 * Time-To-Live (how long a message should exist before it is delivered successfully) in seconds.
+		 * If a message is not delivered successfully within the TTL time, the message is considered expired
+		 * and will be rejected if TTL is supported. Must be between 300 and 259200. Default is 600.
+		 *
+		 * @param ttl The message time-to-live in seconds.
+		 *
+		 * @return This builder.
+		 * @since 8.7.0
+		 */
+		public B ttl(int ttl) {
+			this.ttl = ttl;
 			return (B) this;
 		}
 	}